Return a Result instead of bool from get_interpolant (#88)
Instead of returning a boolean for success upon getting an interplant and throwing an exception on a failure, it will now return a Result. If UNSAT, then the interpolant I should be populated. If SAT then the query was satisfiable and there is no interpolant. Finally, if the result is UNKNOWN there was a failure in the interpolation procedure (for example, no quantifier-free interplant could be found).