Fix for Boolector term iteration (#85)
Previously, the boolector backend made an assumption that term iterators (for getting the children of a term) were only compared if they were iterating over the same node (e.g. only the begin/end iterator for a given `t` not different terms). Comparison between `BoolectorTermIter` used to only compare the `Btor *` instance, and the current position, but not the actual node or children that were being iterated over. This has worked fine so far, but to be more general this PR adds tests for more complicated TermIter comparisons and updates the boolector backend to implement the expected behavior. Here's the new test which the previous term iteration implementation would fail: https://github.com/makaimann/smt-switch/blob/632a43bbf3f89a9ec0393aefccab719e231d8a60/tests/unit/unit-termiter.cpp#L105

I didn't want to have the overhead of finding the right vector of children every time the term is iterated over. It's not completely trivial because the underlying boolector representation of a term doesn't match the smt-switch interface perfectly (for example, negated nodes should have one child in smt-switch). If the children were gathered every time an iterator was returned, it would have to be done for both `begin` and `end` meaning they would be gathered twice for one iteration. Thus, I created a separate method that gathers the children of a `BtorNode` and caches the result.