Fix arity of SMT-LIB core theory functions (#442)
This causes issues in term translation. When we encounter, e.g., a ternary AND, it seems to have incorrect sorts, because the code currently says that AND can only have arity 2. So then we try all kinds of strange transformations to work around sort aliasing in Boolector and other solver idiosyncrasies, when in reality the term is just fine and can be transferred directly.